CAUSEWAY-3815: just renaming packages where actionlinks reside
