+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/components/actionlinks/entityactions/ActionLinksAsButtonInlinePanel.java
@@ -16,7 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.apache.causeway.viewer.wicket.ui.components.actionmenu.entityactions;
+package org.apache.causeway.viewer.wicket.ui.components.actionlinks.entityactions;
 
 import org.apache.causeway.commons.collections.Can;
 import org.apache.causeway.viewer.wicket.ui.components.widgets.actionlink.ActionLink;
