+++ viewers/wicket/ui/src/main/java/org/apache/causeway/viewer/wicket/ui/util/WktDecorators.java
@@ -41,7 +41,7 @@ import org.apache.causeway.viewer.commons.model.decorators.PrototypingDecorator;
 import org.apache.causeway.viewer.commons.model.decorators.TooltipDecorator;
 import org.apache.causeway.viewer.commons.model.decorators.TooltipDecorator.TooltipDecorationModel;
 import org.apache.causeway.viewer.commons.model.layout.UiPlacementDirection;
-import org.apache.causeway.viewer.wicket.ui.components.actionmenu.FontAwesomeBehavior;
+import org.apache.causeway.viewer.wicket.ui.components.actionlinks.FontAwesomeBehavior;
 import org.apache.causeway.viewer.wicket.ui.components.widgets.actionlink.ActionLink;
 import org.apache.causeway.viewer.wicket.ui.util.BootstrapConstants.ButtonSemantics;
